move WindowEvent.tsx, WindowEvent.spec.tsx to utils folder
