+++ redisinsight/ui/src/components/base/utils/WindowEvent.spec.tsx
