+++ redisinsight/ui/src/components/base/utils/WindowEvent.tsx
